requirements.txt
Django==4.1.3
django-jazzmin==2.6.0
django-mptt==0.14.0
django-mptt-admin==2.3.0
